+++ src/estimation/mod.rs
@@ -1,5 +1,6 @@
 pub mod gauss_newton;
 pub(crate) mod hmc;
+pub mod impmap;
 pub mod importance_sampling;
 pub mod inner_optimizer;
 pub mod outer_optimizer;